// src/api/support/types.ts
/* eslint-disable */
// tslint:disable
/**
 * Yeying API
 * 
 *
 * OpenAPI spec version: 1.0
 * Contact: yeying.community@gmail.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator+.
 * https://github.com/karlvr/openapi-generator-plus
 * Do not edit the class manually.
 */

import { Express } from 'express'
import { Api } from '../../models'

export interface SupportApi {
	supportCollect: (request: Api.SupportCollectSupportRequest) => Promise<SupportCollectResponse>
}

export type SupportCollectResponse = SupportCollect200Response | SupportCollectDefaultResponse

export interface SupportCollect200Response {
	status: 200
	body: Api.SupportCollectSupportResponse
	headers?: never
}

export interface SupportCollectDefaultResponse {
	/* Catch-all response */
	status: 'default'
	actualStatus: number
	body: Api.RpcStatus
	headers?: never
}